JADE191

Documentation for JADE191.

JADE191.CityWalk

JADE191.Dijkstra_longest_path

JADE191.buildAdjacencyMatrix

JADE191.calculate_new_distance

JADE191.findSolution

JADE191.getNextVertex

JADE191.getSinglePath!

JADE191.get_car_trace

JADE191.get_paris_layout

JADE191.is_below_line

JADE191.is_out_of_bounds

JADE191.main

JADE191.plot_path

JADE191.read_from_file

JADE191.trial

JADE191.CityWalk — Method

CityWalk(city, adjacencyMatrix, bounding_longitude_data, bounding_latitude_data)

Compute paths in the city for all 8 cars and return a solution.

Returns a Solution object representing the computed paths.

JADE191.Dijkstra_longest_path — Method

Dijkstra_longest_path(adjacencyMatrix, city, start, goal_func, time)

Perform Dijkstra shortest path algorithm.

Finds the path of shortest time connecting start to any junction satisfying goal_func.

JADE191.buildAdjacencyMatrix — Method

buildAdjacencyMatrix(city)

Compute an adjaceny matrix to represent the connections between junctions.

Returns a Dict D where D[A] is a vector of all junctions connected to A by a street.

JADE191.calculate_new_distance — Method

calculate_new_distance(output_so_far, new_path, car, city)

Calculate the new distance traveled by adding a new car path.

JADE191.findSolution — Method

findSolution(start_vertex::Int64,
    total_duration,
    adjacencyMatrix,
    bounding_longitude_data,
    bounding_latitude_data,
    city;
    display_plot=false
)

Compute paths in the city for all 8 cars.

Returns a Vector containing all the paths. We compute the path using a greedy algorithm which splits the map into two regions seperated by the river. At each point of the path we aim to travel as much unique new distance as possible.

JADE191.getNextVertex — Method

getNextVertex(vertex, duration_remaining, visited, adjacencyMatrix,
    bounding_longitude_data, bounding_latitude_data, city,
    should_stay_above)

Compute the next vertex in a given walk.

Returns a Tuple (N, T) where both elements are nothing if the function failed to find a street which is crossable in the amount of time left. If the function is able to find such a street, then N is the next vector in the path and T is the amount of time it takes to travel from the current vertex to N. If shouldstayabove is true, then only vertices located above the dividing line are considered. If shouldstayabove is false, then only vertices located below the dividing line are considered.

JADE191.getSinglePath! — Method

getSinglePath!(adjacencyMatrix, start_vertex, total_duration, visited)

Compute a path in the city which can be traversed in total_duration time.

Returns a Vector representing the path.

JADE191.get_car_trace — Method

get_car_trace(car)

Return the trace of a car. This is used only for plotting.

JADE191.get_paris_layout — Method

get_paris_layout()JADE191sed only for plotting.

JADE191.is_below_line — Method

is_below_line(vertex, bounding_longitude_data, bounding_latitude_data, city)

Calculate if the given vertex is located below or above the seperating curve.

The map of Paris is split along the river that divides the city. Returns true if the given vertex is above that line and false if it is below that line.

JADE191.is_out_of_bounds — Method

is_out_of_bounds(
    vertex, other_vertex, bounding_longitude_data, bounding_latitude_data,
    city
)

Calculate if a vertex and another vertex are within according to the division.

JADE191.main — Method

main()

Perform a single run of the path-finding algorithm.

JADE191.plot_path — Method

plot_path(p, car, path, city)

Plot the path of a car traveling through the city.

JADE191.read_from_file — Method

read_from_file(filename)

Read the coordinates defining the splitting line through the river.

JADE191.trial — Method

trial(
    adjacencyMatrix,
    start_vertex,
    total_duration,
    bounding_longitude_data,
    bounding_latitude_data,
    city,
    expected_diff_per_car,
    still_apply_expectation_max_value,
    max_attempts,
    num_cars,
    trial_num,
    should_stay_above,
)

Perform a single trial of finding four cars to map out a region of Paris.