JADE191

Documentation for JADE191.

JADE191.CityWalk

JADE191.buildAdjacencyMatrix

JADE191.findExtremelyNaiveSolution

JADE191.getNextVertex

JADE191.getSinglePath!

JADE191.main

JADE191.CityWalk — Method

CityWalk(city, adjacencyMatrix)

Compute paths in the city for all 8 cars and return a solution.

Returns a Solution object representing the computed paths.

JADE191.buildAdjacencyMatrix — Method

buildAdjacencyMatrix(city)

Compute an adjaceny matrix to represent the connections between junctions.

Returns a Dict D where D[A] is a vector of all junctions connected to A by a street.

JADE191.findExtremelyNaiveSolution — Method

findExtremelyNaiveSolution(start_vertex, total_duration, adjacencyMatrix)

Compute paths in the city for all 8 cars.

Returns a Vector containing all the paths. We compute the path using a greedy algorithm, where at each point of the path we aim to travel as much unique new distance as possible.

JADE191.getNextVertex — Method

getNextVertex(vertex, duration_remaining, visited, adjacencyMatrix)

Compute the next vertex in a given walk.

Returns a Tuple (N, T) where both elements are nothing if the function failed to find a street which is crossable in the amount of time left. If the function is able to find such a street, then N is the next vector in the path and T is the amount of time it takes to travel from the current vertex to N.

JADE191.getSinglePath! — Method

getSinglePath!(adjacencyMatrix, start_vertex, total_duration, visited)

Compute a path in the city which can be traversed in total_duration time.

Returns a Vector representing the path.

JADE191.main — Method

main()

Loads the city data, computes solutions, and stores them.